Corrigé un problème qui générait un warning de parsing d'options. Modifié l'image de la graine de rosier.
git-svn-id: file:///Users/matti/Desktop/tmp-1/flore/svn/trunk@182 0452f256-4c2a-4938-993f-bf7dc213af01

+++ flore/src/xml/GlobalProperties.java
@@ -116,7 +116,7 @@ public class GlobalProperties {
 	public static Double getDouble(String key) {
 		init();
 		try {
-			return new Integer(props.getProperty(key)) / STORAGE_NUMBERS_COEF;
+			return new Double(props.getProperty(key)) / STORAGE_NUMBERS_COEF;
 		} catch (NumberFormatException e) {
 			System.err.println("Couldn't parse option \"" + key + "\" as double ! Returned default value " + DEFAULT_DOUBLE_VALUE + " instead.");
 			return DEFAULT_DOUBLE_VALUE;
